v2019_update_license_year

+++ src/PdfSharp/Internal/Diagnostics.cs
@@ -3,7 +3,7 @@
 // Authors:
 //   Stefan Lange
 //
-// Copyright (c) 2005-2016 empira Software GmbH, Cologne Area (Germany)
+// Copyright (c) 2005-2019 empira Software GmbH, Cologne Area (Germany)
 //
 // http://www.pdfsharp.com
 // http://sourceforge.net/projects/pdfsharp
@@ -110,7 +110,7 @@ namespace PdfSharp.Internal
         {
             string message = string.Format(CultureInfo.InvariantCulture,
                 "Unexpected character '0x{0:x4}' in content stream. The stream may be corrupted or the feature is not implemented. " +
-                "If you think this is a bug in PDFsharp, please send us your PDF file.", ch);
+                "If you think this is a bug in PDFsharp, please send us your PDF file.", (int)ch);
             ThrowContentReaderException(message);
         }
     }
